+++ rpxy-bin/src/error.rs
@@ -1,1 +1,2 @@
+#[allow(unused)]
 pub use anyhow::{anyhow, bail, ensure, Context};