+++ rpxy-bin/src/log.rs
@@ -1,24 +1,44 @@
+use std::str::FromStr;
+use tracing_subscriber::{fmt, prelude::*};
+
+#[allow(unused)]
 pub use tracing::{debug, error, info, warn};
 
+/// Initialize the logger with the RUST_LOG environment variable.
 pub fn init_logger() {
-  use tracing_subscriber::{fmt, prelude::*, EnvFilter};
+  let level_string = std::env::var("RUST_LOG").unwrap_or_else(|_| "info".to_string());
+  let level = tracing::Level::from_str(level_string.as_str()).unwrap_or(tracing::Level::INFO);
 
-  let format_layer = fmt::layer()
-    .with_line_number(false)
-    .with_thread_ids(false)
-    .with_target(false)
-    .with_thread_names(true)
-    .with_target(true)
-    .with_level(true)
-    .compact();
-
-  // This limits the logger to emits only rpxy crate
-  // let level_string = std::env::var(EnvFilter::DEFAULT_ENV).unwrap_or_else(|_| "info".to_string());
-  // let filter_layer = EnvFilter::new(format!("{}={}", env!("CARGO_PKG_NAME"), level_string));
-  let filter_layer = EnvFilter::from_default_env();
-
-  tracing_subscriber::registry()
-    .with(format_layer)
-    .with(filter_layer)
-    .init();
+  // This limits the logger to emits only this crate with any level above RUST_LOG, for included crates it will emit only ERROR (in prod)/INFO (in dev) or above level.
+  let stdio_layer = fmt::layer().with_level(true).with_thread_ids(false);
+  if level <= tracing::Level::INFO {
+    // in normal deployment environment
+    let stdio_layer = stdio_layer
+      .with_target(false)
+      .compact()
+      .with_filter(tracing_subscriber::filter::filter_fn(move |metadata| {
+        (metadata
+          .target()
+          .starts_with(env!("CARGO_PKG_NAME").replace('-', "_").as_str())
+          && metadata.level() <= &level)
+          || metadata.level() <= &tracing::Level::ERROR.min(level)
+      }));
+    tracing_subscriber::registry().with(stdio_layer).init();
+  } else {
+    // debugging
+    let stdio_layer = stdio_layer
+      .with_line_number(true)
+      .with_target(true)
+      .with_thread_names(true)
+      .with_target(true)
+      .compact()
+      .with_filter(tracing_subscriber::filter::filter_fn(move |metadata| {
+        (metadata
+          .target()
+          .starts_with(env!("CARGO_PKG_NAME").replace('-', "_").as_str())
+          && metadata.level() <= &level)
+          || metadata.level() <= &tracing::Level::INFO.min(level)
+      }));
+    tracing_subscriber::registry().with(stdio_layer).init();
+  };
 }