+++ rpxy-lib/src/message_handler/http_result.rs
@@ -8,8 +8,8 @@ pub(crate) type HttpResult<T> = std::result::Result<T, HttpError>;
 /// Describes things that can go wrong in the forwarder
 #[derive(Debug, Error)]
 pub enum HttpError {
-  #[error("No host is give nin request header")]
-  NoHostInRequestHeader,
+  // #[error("No host is give in request header")]
+  // NoHostInRequestHeader,
   #[error("Invalid host in request header")]
   InvalidHostInRequestHeader,
   #[error("SNI and Host header mismatch")]
@@ -32,11 +32,10 @@ pub enum HttpError {
 
   #[error("Failed to upgrade connection: {0}")]
   FailedToUpgrade(String),
-  #[error("Request does not have an upgrade extension")]
-  NoUpgradeExtensionInRequest,
-  #[error("Response does not have an upgrade extension")]
-  NoUpgradeExtensionInResponse,
-
+  // #[error("Request does not have an upgrade extension")]
+  // NoUpgradeExtensionInRequest,
+  // #[error("Response does not have an upgrade extension")]
+  // NoUpgradeExtensionInResponse,
   #[error(transparent)]
   Other(#[from] anyhow::Error),
 }
@@ -44,7 +43,7 @@ pub enum HttpError {
 impl From<HttpError> for StatusCode {
   fn from(e: HttpError) -> StatusCode {
     match e {
-      HttpError::NoHostInRequestHeader => StatusCode::BAD_REQUEST,
+      // HttpError::NoHostInRequestHeader => StatusCode::BAD_REQUEST,
       HttpError::InvalidHostInRequestHeader => StatusCode::BAD_REQUEST,
       HttpError::SniHostInconsistency => StatusCode::MISDIRECTED_REQUEST,
       HttpError::NoMatchingBackendApp => StatusCode::SERVICE_UNAVAILABLE,
@@ -54,8 +53,8 @@ impl From<HttpError> for StatusCode {
       HttpError::FailedToAddSetCookeInResponse(_) => StatusCode::INTERNAL_SERVER_ERROR,
       HttpError::FailedToGenerateDownstreamResponse(_) => StatusCode::INTERNAL_SERVER_ERROR,
       HttpError::FailedToUpgrade(_) => StatusCode::INTERNAL_SERVER_ERROR,
-      HttpError::NoUpgradeExtensionInRequest => StatusCode::BAD_REQUEST,
-      HttpError::NoUpgradeExtensionInResponse => StatusCode::BAD_GATEWAY,
+      // HttpError::NoUpgradeExtensionInRequest => StatusCode::BAD_REQUEST,
+      // HttpError::NoUpgradeExtensionInResponse => StatusCode::BAD_GATEWAY,
       _ => StatusCode::INTERNAL_SERVER_ERROR,
     }
   }