+++ CHANGELOG.md
@@ -1,12 +1,25 @@
 # CHANGELOG
 
-## 0.7.0  (unreleased)
+## 0.8.0 (Unreleased)
+
+## 0.7.1 -- 0.7.3
+
+- deps and patches
+
+## 0.7.0
+
+### Important Changes
 
 - Breaking: `hyper`-1.0 for both server and client modules.
-- Breaking: Remove `override_host` option in upstream options. Add a reverse option, i.e., `disable_override_host`. That is, `rpxy` always override the host header by the upstream hostname by default.
+- Breaking: Remove `override_host` option in upstream options. Add a reverse option, i.e., `keep_original_host`, and the similar option `set_upstream_host`. While `keep_original_host` can be explicitly specified, `rpxy` keeps the original `host` given by the incoming request by default. Then, the original `host` header is maintained or added from the value of url request line. If `host` header needs to be overridden with the upstream host name (backend uri's host name), `set_upstream_host` has to be set. If both of `set_upstream_host` and `keep_original_host` are set, `keep_original_host` is prioritized since it is explicitly specified.
 - Breaking: Introduced `native-tls-backend` feature to use the native TLS engine to access backend applications.
+- Breaking: Changed the policy of the default cert store from `webpki` to the system-native store. Thus we terminated the feature `native-roots` and introduced `webpki-roots` feature to use `webpki` root cert store.
+
+### Improvement
+
 - Redesigned: Cache structure is totally redesigned with more memory-efficient way to read from cache file, and more secure way to strongly bind memory-objects with files with hash values.
 - Redesigned: HTTP body handling flow is also redesigned with more memory-and-time efficient techniques without putting the whole objects on memory by using `futures::stream::Stream` and `futures::channel::mpsc`
+- Feat: Allow to disable/enable forced-connection-timeout regardless of connection status (idle or not). [default: disabled]
 - Refactor: lots of minor improvements
 
 ## 0.6.2